feat(go): rename wd_info_shot -> wd_short_info
change ParseWoodpeckerInfo2Shot -> ParseWoodpeckerInfo2Short

BREAKING CHANGE: must change package wd_info_shot -> wd_short_info and method
ParseWoodpeckerInfo2Shot -> ParseWoodpeckerInfo2Short

fix #34
